workers/requirements.txt
celery[redis]>=5.0
redis>=4.0
ffmpeg-python>=0.2.0
openai-whisper>=20231117
boto3>=1.20
